libcrux_intrinsics::avx2

Function _xrstors64
·

pub unsafe fn _xrstors64(mem_addr: *const u8, rs_mask: u64)

Performs a full or partial restore of the enabled processor states using the state information stored in memory at mem_addr.

xrstors differs from xrstor in that it can restore state components corresponding to bits set in the IA32_XSS MSR; xrstors cannot restore from an xsave area in which the extended region is in the standard form. State is restored based on bits [62:0] in rs_mask, XCR0, and mem_addr.HEADER.XSTATE_BV. mem_addr must be aligned on a 64-byte boundary.

Intel’s documentation